Continuação da revisão da arquitetura do web: Adição final dos diferentes tipos de pastas criadas por  pelo menos uma funcionalidade, podendo usar de exemplo para o restante dos códigos e novas implementações e funções a serem feitas. As pastas criadas foram de constants e types, onde a pasta de constants deve manter a utilização de consts criados em qualquer código no sistema, enquanto a pasta de types será criada para cada função ou utilizada na shared para types e interfaces usadas nos códigos. Por fim foi feito três documentos detalhando a mudança de arquitetura feita em uma visão geral, na visão da arquitetura de web, e na visão da arquitetura do mobile
