+++ apps/web/package.json
@@ -16,7 +16,7 @@
     "bootstrap": "^5.3.8",
     "bootstrap-icons": "^1.13.1",
     "jwt-decode": "^4.0.0",
-    "lucide-react": "^1.7.0",
+    "lucide-react": "^1.25.0",
     "react": "19.1.0",
     "react-bootstrap": "^2.10.10",
     "react-dom": "19.1.0",
