+++ apps/web/src/contexts/authContext.tsx
@@ -1,34 +1,7 @@
 import { createContext, useEffect, useState } from "react";
 import { loginRequest } from "@/features/auth/services/authService";
-
-// ─── Mapeamento enum numérico do backend → string ────────────────────────────
-// Domain.Enums.TipoUsuario: Instrutor=1, Aluno=2, Gestor=3, Administrador=4
-const tipoUsuarioMap: Record<number, string> = {
-  1: "Instrutor",
-  2: "Aluno",
-  3: "Gestor",
-  4: "Administrador",
-};
-
-// ─── Types ────────────────────────────────────────────────────────────────────
-
-export interface User {
-  Id: string;
-  TipoUsuario: string;      // "Gestor" | "Instrutor" | "Aluno" | "Administrador"
-  IdAcademia: string | null; // preenchido após atualização do backend
-  Flag: boolean;
-  Nome: string | null;       // null até endpoint de perfil ser implementado
-  AvatarUrl: string | null;  // null até endpoint de perfil ser implementado
-}
-
-export interface AuthContextType {
-  user: User | null;
-  isAuthenticated: boolean;
-  loading: boolean;
-  login: (email: string, password: string) => Promise<User>;
-  logout: () => void;
-  updateUser: (user: User) => void;
-}
+import { tipoUsuarioMap } from "@/shared/constants/Roles";
+import type { User, AuthContextType } from "@/shared/types/auth.ts";
 
 // ─── Context ──────────────────────────────────────────────────────────────────
