+++ apps/web/src/features/auth/components/login/LoginForm.tsx
@@ -1,34 +1,9 @@
 import { useState } from "react";
-import { useAuth } from "@/shared/hooks/useAuth";
 import { useNavigate } from "react-router-dom";
+import { Mail, Lock, Check } from "lucide-react";
+import { useAuth } from "@/shared/hooks/useAuth";
 import { InputField } from "@/shared/components/inputs/InputGenericoModerno";
 import { RoleRoutes } from "@/shared/constants/Roles";
-
-function EnvelopeIcon() {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <rect x="2" y="4" width="20" height="16" rx="2" />
-      <path d="m22 7-8.97 5.7a1.94 1.94 0 0 1-2.06 0L2 7" />
-    </svg>
-  );
-}
-
-function LockIcon() {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <rect x="3" y="11" width="18" height="11" rx="2" />
-      <path d="M7 11V7a5 5 0 0 1 10 0v4" />
-    </svg>
-  );
-}
-
-function CheckIcon() {
-  return (
-    <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round">
-      <polyline points="20 6 9 17 4 12" />
-    </svg>
-  );
-}
 
 export default function LoginForm() {
   const { login } = useAuth();
@@ -93,7 +68,7 @@ export default function LoginForm() {
           type="email"
           value={email}
           onChange={(e) => setEmail(e.target.value)}
-          icon={<EnvelopeIcon />}
+          icon={<Mail/>}
           error={emailError}
         />
 
@@ -102,7 +77,7 @@ export default function LoginForm() {
           label="Senha"
           value={password}
           onChange={(e) => setPassword(e.target.value)}
-          icon={<LockIcon />}
+          icon={<Lock/>}
           error={passwordError}
           isPassword
         />
@@ -123,7 +98,7 @@ export default function LoginForm() {
                 ? "border-[#18181b] dark:border-[#e4e4e7] bg-[#18181b] dark:bg-[#e4e4e7] text-white dark:text-[#18181b]"
                 : "border-[#d4d4d8] dark:border-[rgba(255,255,255,0.2)] bg-white dark:bg-[#2c2c30]"}
             `}>
-              {rememberMe && <CheckIcon />}
+              {rememberMe && <Check/>}
             </span>
             <span className="font-semibold text-[#52525c] dark:text-[#a1a1aa]">
               Lembrar-me
