+++ apps/web/src/features/auth/services/authService.tsx
@@ -1,16 +1,6 @@
 import { api } from "@/shared/services/apiURL";
-
+import type { LoginApiResponse } from "@/features/auth/type/index.ts";
 // ─── Types ────────────────────────────────────────────────────────────────────
-
-interface LoginApiResponse {
-  IdUsuario: string;
-  TipoUsuario: number;   // enum numérico: 1=Instrutor, 2=Aluno, 3=Gestor, 4=Administrador
-  IdAcademia: string;    // adicionado no backend — Guid da academia do gestor
-  Flag: boolean;
-  Token: string;
-  RefreshToken: string;
-  Status: { Message: string; Code: number; Sucess: boolean };
-}
 
 export interface UserProfileResponse {
   Nome: string;
