+++ apps/web/src/shared/constants/Roles.ts
@@ -1,5 +1,12 @@
 import { ConstRoutes } from "./Routes";
 
+export const tipoUsuarioMap: Record<number, string> = {
+  1: "Instrutor",
+  2: "Aluno",
+  3: "Gestor",
+  4: "Administrador",
+}
+
 export const RoleRoutes: Record<string, string> = {
   Instrutor:     ConstRoutes.INSTRUTOR,
   Aluno:         ConstRoutes.ALUNO,
