+++ apps/web/src/shared/hooks/useAuth.tsx
@@ -4,5 +4,5 @@ import { AuthContext } from "@/contexts/AuthContext";
 export function useAuth() {
   const context = useContext(AuthContext);
   if (!context) throw new Error("useAuth deve ser usado dentro do AuthProvider");
-  return context;
+    return context;
 }
